Update supported clang/llvm for pxbind (#153)
* Update clang/llvm support

* Regenerate files

* Update x86_64-unknown-linux

* Update rigid dynamic flags

* Fix formatting

* Update x86_64-apple-darwin

* Update x86_64-pc-windows-msvc

* Remove unnecessery llvm library

+++ physx/src/rigid_dynamic.rs
@@ -45,13 +45,11 @@ impl PxFlags for RigidDynamicLockFlags {
     type Target = PxRigidDynamicLockFlags;
 
     fn into_px(self) -> Self::Target {
-        PxRigidDynamicLockFlags {
-            mBits: self.bits() as u16,
-        }
+        PxRigidDynamicLockFlags { mBits: self.bits() }
     }
 
     fn from_px(flags: Self::Target) -> Self {
-        unsafe { BitFlags::from_bits_unchecked(flags.mBits as u8) }
+        unsafe { BitFlags::from_bits_unchecked(flags.mBits) }
     }
 }
